estate_planner/requirements.txt
streamlit>=1.35.0
anthropic>=0.30.0
fpdf2>=2.7.0
jinja2>=3.1.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
